backend/requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-auditlog==3.3.0
django-cors-headers==4.9.0
django-jsonfield==1.4.1
django-rest-durin==1.1.0
djangorestframework==3.16.1
gunicorn==23.0.0
humanize==4.14.0
idna==3.11
packaging==25.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
